xwrap-ldplugin: more correct version
This fixes the xwrapping case where a symbol is defined in an external
DSO, in which case we revert to --wrap.

The code is rather nasty... it parses the linker command line in order
to identify, early, which symbols are to be locally defined in the link,
then restarts if --wrap is missing for any of them. If we the plugin API
let us mutate the set of --wrap options in effect, this would not be
necessary.